+++ js/jquery/components/Modal.js
@@ -267,7 +267,7 @@
         template: '<div class="modal">' +
             '<div class="modal-handle">' +
                 '<div class="modal-inner"></div>' +
-                '<a href="javascript:;" class="modal-close modal-event-close"><span class="x"></span></a>' +
+                '<button type="button" class="modal-close modal-event-close"><span class="x"></span></button>' +
             '</div>' +
         '</div>'
     });